pyproject.toml
[project]
name = "llm-endtoend-proje"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-google-genai==2.1.8",
    "langchain-qdrant==0.2.1",
    "langchain-tavily>=0.2.12",
    "langgraph==0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.6",
    "pypdf==6.0.0",
    "pysqlite3-binary>=0.5.4",
    "python-dotenv==1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "unstructured>=0.18.18",
    "uvicorn>=0.38.0",
]
